apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pydantic>=2,<3
pydantic-settings>=2,<3

SQLAlchemy>=2.0
psycopg2-binary>=2.9
email-validator>=2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

Pillow>=10.0
python-multipart>=0.0.9

alembic>=1.13
httpx>=0.27
google-cloud-vision
